trilby-cli: refactor install config, expose Configuration and Host modules
