+++ trilby-cli/src/Trilby/Host.hs
@@ -20,6 +20,16 @@ instance Show Host where
     show Host{username = Nothing, ..} = Text.unpack hostname
     show Host{username = Just username, ..} = Text.unpack $ username <> "@" <> hostname
 
+hostname :: Host -> App Text
+hostname Localhost = view #hostname
+hostname Host{..} = pure hostname
+
+canonicalHost :: Host -> App Host
+canonicalHost host = do
+    localhostHostnames <- view #hostname <&> (: ["localhost", "127.0.0.1", "::1"])
+    isLocalhost <- hostname host <&> (`elem` localhostHostnames)
+    pure $ if isLocalhost then Localhost else host
+
 -- | Execute a command over SSH, if given a remote host.
 ssh :: Host -> (NonEmpty Text -> App ()) -> NonEmpty Text -> App ()
 ssh Localhost c t = c t
