+++ trilby-cli/src/Trilby/Update.hs
@@ -1,30 +1,18 @@
-{-# OPTIONS_GHC -Wno-name-shadowing #-}
-
 module Trilby.Update (update) where
 
 import Data.List.NonEmpty.Extra qualified as NonEmpty
-import Data.Text qualified as Text
 import Trilby.HNix (FileOrFlake (..), FlakeRef (..), nixBuild, writeNixFile)
 import Trilby.Update.Options
 import Turtle (readlink, (</>))
 import Trilby.Host
+import Trilby.Configuration (Configuration(..))
+import Trilby.Configuration qualified as Configuration
 import Prelude
 
 update :: UpdateOpts Maybe -> App ()
 update (askOpts -> opts) = do
     whenM opts.flakeUpdate . inDir "/etc/trilby" $ rawCmd_ ["nix", "flake", "update", "--accept-flake-config"]
-    hostname <- Text.strip <$> cmd ["hostnamectl", "hostname"]
-    let canonicalHost :: Host -> Host
-        canonicalHost Localhost = Localhost
-        canonicalHost host@Host{hostname = h}
-            | h `elem` hostname :| ["localhost", "127.0.0.1", "::1"] = Localhost
-            | otherwise = host
-    let configuration :: Host -> Configuration
-        configuration =
-            canonicalHost >>> \case
-                Localhost -> Configuration{name = hostname, host = Localhost}
-                host@Host{hostname} -> Configuration{name = hostname, host}
-    configurations <- configuration <$$> opts.hosts <&> NonEmpty.nubOrd
+    configurations <- mapM Configuration.fromHost . NonEmpty.nubOrd =<< opts.hosts
     configurationResults <- buildConfigurations configurations
     for_ configurationResults $ \(Configuration{..}, resultPath) -> do
         unless (host == Localhost) $ copyClosure host resultPath
@@ -38,9 +26,6 @@ update (askOpts -> opts) = do
                 whenM reboot $ ssh host cmd_ ["systemctl", "reboot"]
             Test -> perform ConfigTest
             NoAction -> pure ()
-
-data Configuration = Configuration {name :: Text, host :: Host}
-    deriving stock (Generic, Eq, Ord)
 
 buildConfiguration :: Configuration -> App (Configuration, FilePath)
 buildConfiguration c = (c,) <$> nixBuild f
